:heavy_plus_sign: Add Files class
