+++ redcap/project.py
@@ -6,6 +6,7 @@ from io import StringIO
 
 from redcap.request import RedcapError
 from redcap.methods.field_names import FieldNames
+from redcap.methods.files import Files
 from redcap.methods.metadata import Metadata
 
 
@@ -21,7 +22,7 @@ __copyright__ = "2014, Vanderbilt University"
 # pylint: disable=consider-using-f-string
 # pylint: disable=consider-using-generator
 # pylint: disable=use-dict-literal
-class Project(FieldNames, Metadata):
+class Project(FieldNames, Files, Metadata):
     """Main class for interacting with REDCap projects"""
 
     def export_fem(self, arms=None, format="json", df_kwargs=None):
@@ -251,19 +252,6 @@ class Project(FieldNames, Metadata):
             field_name, "text_validation_type_or_show_slider_number"
         )
 
-    def _meta_metadata(self, field, key):
-        """Return the value for key for the field in the metadata"""
-        metadata_field = ""
-        try:
-            metadata_field = str(
-                [f[key] for f in self.metadata if f["field_name"] == field][0]
-            )
-        except IndexError:
-            print("%s not in metadata field:%s" % (key, field))
-            return metadata_field
-        else:
-            return metadata_field
-
     def backfill_fields(self, fields, forms):
         """
         Properly backfill fields to explicitly request specific
@@ -357,164 +345,6 @@ class Project(FieldNames, Metadata):
         if "error" in response:
             raise RedcapError(str(response))
         return response
-
-    def export_file(
-        self, record, field, event=None, return_format="json", repeat_instance=None
-    ):
-        """
-        Export the contents of a file stored for a particular record
-
-        Notes
-        -----
-        Unlike other export methods, this works on a single record.
-
-        Parameters
-        ----------
-        record : str
-            record ID
-        field : str
-            field name containing the file to be exported.
-        event: str
-            for longitudinal projects, specify the unique event here
-        return_format: ('json'), 'csv', 'xml'
-            format of error message
-        repeat_instance: (None),str,int
-
-        Returns
-        -------
-        content : bytes
-            content of the file
-        content_map : dict
-            content-type dictionary
-        """
-        self._check_file_field(field)
-        # load up payload
-        payload = self._basepl(content="file", format=return_format)
-        # there's no format field in this call
-        del payload["format"]
-        payload["returnFormat"] = return_format
-        payload["action"] = "export"
-        payload["field"] = field
-        payload["record"] = record
-        if event:
-            payload["event"] = event
-        if repeat_instance:
-            payload["repeat_instance"] = str(repeat_instance)
-        content, headers = self._call_api(payload, "exp_file")
-        # REDCap adds some useful things in content-type
-        content_map = {}
-        if "content-type" in headers:
-            splat = [
-                key_values.strip() for key_values in headers["content-type"].split(";")
-            ]
-            key_values = [
-                (key_values.split("=")[0], key_values.split("=")[1].replace('"', ""))
-                for key_values in splat
-                if "=" in key_values
-            ]
-            content_map = dict(key_values)
-
-        return content, content_map
-
-    def import_file(
-        self,
-        record,
-        field,
-        fname,
-        fobj,
-        event=None,
-        repeat_instance=None,
-        return_format="json",
-    ):
-        """
-        Import the contents of a file represented by fobj to a
-        particular records field
-
-        Parameters
-        ----------
-        record : str
-            record ID
-        field : str
-            field name where the file will go
-        fname : str
-            file name visible in REDCap UI
-        fobj : file object
-            file object as returned by `open`
-        event : str
-            for longitudinal projects, specify the unique event here
-        repeat_instance : int
-            (only for projects with repeating instruments/events)
-            The repeat instance number of the repeating event (if longitudinal)
-            or the repeating instrument (if classic or longitudinal).
-        return_format : ('json'), 'csv', 'xml'
-            format of error message
-
-        Returns
-        -------
-        response :
-            response from server as specified by ``return_format``
-        """
-        self._check_file_field(field)
-        # load up payload
-        payload = self._basepl(content="file", format=return_format)
-        # no format in this call
-        del payload["format"]
-        payload["returnFormat"] = return_format
-        payload["action"] = "import"
-        payload["field"] = field
-        payload["record"] = record
-        if event:
-            payload["event"] = event
-        if repeat_instance:
-            payload["repeat_instance"] = repeat_instance
-        file_kwargs = {"files": {"file": (fname, fobj)}}
-        return self._call_api(payload, "imp_file", **file_kwargs)[0]
-
-    def delete_file(self, record, field, return_format="json", event=None):
-        """
-        Delete a file from REDCap
-
-        Notes
-        -----
-        There is no undo button to this.
-
-        Parameters
-        ----------
-        record : str
-            record ID
-        field : str
-            field name
-        return_format : (``'json'``), ``'csv'``, ``'xml'``
-            return format for error message
-        event : str
-            If longitudinal project, event to delete file from
-
-        Returns
-        -------
-        response : dict, str
-            response from REDCap after deleting file
-        """
-        self._check_file_field(field)
-        # Load up payload
-        payload = self._basepl(content="file", format=return_format)
-        del payload["format"]
-        payload["returnFormat"] = return_format
-        payload["action"] = "delete"
-        payload["record"] = record
-        payload["field"] = field
-        if event:
-            payload["event"] = event
-        return self._call_api(payload, "del_file")[0]
-
-    def _check_file_field(self, field):
-        """Check that field exists and is a file field"""
-        is_field = field in self.field_names
-        is_file = self._meta_metadata(field, "field_type") == "file"
-        if not (is_field and is_file):
-            msg = "'%s' is not a field or not a 'file' field" % field
-            raise ValueError(msg)
-
-        return True
 
     def export_users(self, format="json"):
         """
